Lab3-Parallelization_Analyst/PhoneAnalysisAgent/requirements.txt
streamlit
snowflake-connector-python
python-dotenv
langchain
langchain-community
langchain-core
requests